pyproject.toml
[tool.poetry]
name = "air-template"
version = "1.0.0"
description = "AIR functions"
authors = ["Arun Kaashyap Arunachalam <arun.arunachalam@cognite.com>"]

[tool.black]
line-length = 120
target_version = ['py37']
include = '\.py$'

[tool.isort]
line_length=120                # corresponds to -w  flag
multi_line_output=3            # corresponds to -m  flag
include_trailing_comma=true    # corresponds to -tc flag
skip_glob = '^((?!py$).)*$'    # this makes sort all Python files
known_third_party = ["cognite"]


[tool.poetry.dependencies]
python = "^3.7"
cognite-sdk = "^2"
pyjwt = "^1.7.1"
cognite-sdk-experimental = "^0"
pytest-custom_exit_code = "^0.3.0"
gitpython = "^3.1.1"
PyGithub = "^1.51"
cognite-air-sdk = "^3"
cognite-air-workflow = "^2"


[tool.poetry.dev-dependencies]
isort = "5.6.4"
pytest = "^6.0.1"
pytest-cov = "^2.10.1"
pre-commit = "^2.7.1"
black = "20.8b1"
mypy = "0.790"
flake8 = "3.8.4"
bandit = "^1.6.2"



[virtualenvs]
create = true
in-project = true


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
